backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
pinecone==5.1.0
sentence-transformers==3.0.1
numpy==1.26.4
pydantic==2.8.2
